--- a/addons/account/__manifest__.py
+++ b/addons/account/__manifest__.py
@@ -110,6 +110,7 @@ You could use this simplified accounting in case you work with an (external) acc
             'account/static/src/js/bills_upload.js',
             'account/static/src/js/account_selection.js',
             'account/static/src/js/open_move_widget.js',
+            'account/static/src/js/account_move_form.js',
         ],
         'web.assets_frontend': [
             'account/static/src/js/account_portal_sidebar.js',