--- a/addons/account/models/company.py
+++ b/addons/account/models/company.py
@@ -9,7 +9,6 @@ from odoo.exceptions import ValidationError, UserError, RedirectWarning
 from odoo.tools.mail import is_html_empty
 from odoo.tools.misc import format_date
 from odoo.tools.float_utils import float_round, float_is_zero
-from odoo.tests.common import Form
 
 
 MONTH_SELECTION = [