@robodoo 2022-07-15
[Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/96134).