@william-andre 2022-07-25
@odoo/rd-security 
Hello, I am already asking for a review so that you will be ready when we want to merge as nothing should "change" security wise anymore.
There is only one warning

> [odoo/addons/account/models/account_move.py:2467](https://github.com/odoo/odoo/blob/7e12acb7d16b86dc0befdc02286b6fd6de5ff8ca/addons/account/models/account_move.py#L2467)
> ```python
> ref_function = getattr(self, > f'_get_invoice_reference_{self.journal_id.invoice_reference_model}_{self.journal_id.invoice_reference_type}')
> ```
> Line may contain security risk: getattr. Please contact security team for advanced review (you can ping @odoo/rd-security)

The code is only moved. It should also be safe because it is getting values that are built from `fields.Selection` values