@william-andre 2022-07-26
Hello @legalsylvain 
Thank you for your review, however I am not sure how I could improve the commit message.
It describes the commit really well, on top of the diff being very easy to read, and short.

You should note that the diff is `+6,500 −6,887` at the time of writing this, which is irrevocable proof that the code is now better.

I hope everything is now clearer.

All the best,
William.