@jco-odoo 2022-07-26
[IMP] account: onchange -> compute would already explain something