@william-andre 2022-07-26
@legalsylvain I hope this is a bit better. If you have suggestions, feel free.