@william-andre 2022-07-26
@odoo/rd-security This is now ready for review, more explanation about the changes here: https://github.com/odoo/odoo/pull/96134#issuecomment-1193906077
It shouldn't be an issue for you :blush: