@legalsylvain 2022-07-27
> @legalsylvain I hope this is a bit better. If you have suggestions, feel free.

I was more used to the naming "-pocalypse", but "-yeaah" is OK for me.
Joke aside, thanks for the detail ! 

Nice work. onchange -> compute function.