@robodoo 2022-07-28
@qdp-odoo, you may want to rebuild or fix this PR as it has failed CI.