@rvalyi 2022-08-02
Hello @william-andre @qdp-odoo @xmo-odoo 
for localizations we often need to override many of the core Odoo accounting methods. But we can only do it efficiently if these methods aren't too fat... From a quick scan I spotted the following very large methods, it would be very cool if you could extract submethods from them where it makes sense (specially before creating stuff in the database for instance). Here is the list of the methods and their number of lines in this new version:

**account.move:**

- _post 150
- _sync_dynamic_line 105
- _compute_payment_state 83
- _compute_name 76
- write 68 (not so large but very often overriden)

**account.move.line:**

- _compute_account_id 107
- write   94
- _prepare_reconciliation_single_partial 265
- _add_exchange_difference_cash_basis_vals 170
- reconcile 145 (large but not problematic)

cc @pedrobaeza @alexis-via @sebastienbeau @renatonlima