@william-andre 2022-08-02
@rvalyi Apart from `_compute_account_id` and `_sync_dynamic_line`, all of these were already there in the older version. And `_compute_account_id` is extended a few times in the standard code, always easily and without any "efficiency" issue.
Regarding `_sync_dynamic_line`, I would highly recommend against extending that, if needed, you can just change the parameters.

The other methods are also extended quite a few times in standard code without any issues.
I actually removed a few of those big methods.