@rvalyi 2022-08-02
Thank you @william-andre for clarifying this. Now if you could split just one method, please focus on the _post method. This one sucks, specially here in Brazil were we should take the electronic invoicing transmission state into account before actually posting the move...