@lem8r 2022-08-02
2 cents regarding localization and inheritance.

`_get_price_unit()` from `purchase_stock/models/stock.py`

It converts price unit to company's currency at today's rate. At the moment when picking is validated.
This affects the valuation of a product.

Here is the localization case:
We are importing products from foreign country. PO has prices in foreign currency and we must valuate the products using exchange rate that was on a day when it crossed the border (customs).

Currently it can be achieved by copy-pasting original method.

In general - it would be great to have an option to set accounting date on a picking itself. Same feature already exists on inventory adjustment.