@hoangtrann 2022-11-06
I'm not sure where to ask this question, but as this PR improved a lot for accounting, this sounds like the right place to me...

https://github.com/odoo/odoo/blob/abe4efc0beb949a2add49380e9cb0fdf1fcd7ce3/addons/account/models/account_move.py#L854-L866

I wonder if these was any discussion, or any preferred direction that we could improve this nested dependent? Previously, anytime I work on the performance issues related to `account.move`, it usually leads to these `matched_credit_ids` and `matched_debit_ids`.