@william-andre 2022-11-06
@hoangtrann 
I'm not sure how this can be improved and in which flow these dependencies cause noticeable performance issues.
Do you have a question or a suggestion?